requirements.txt
annotated-types==0.7.0
anyio==4.5.2
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
cliapp==1.20180812.1
Click==7.0
cloud-init==24.4.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info==0.23+ubuntu1.1
dnspython==2.6.1
email-validator==2.2.0
entrypoints==0.3
exceptiongroup==1.2.2
Faker==35.2.2
fastapi==0.115.8
fastapi-cli==0.0.7
h11==0.14.0
httpcore==1.0.7
httplib2==0.14.0
httptools==0.6.4
httpx==0.28.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
jinja2==3.1.5
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
kafka-python==2.1.5
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Markdown==3.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.24.4
oauthlib==3.1.0
packaging==20.3
pandas==2.0.3
pexpect==4.6.0
py4j==0.10.9.7
pyarrow==17.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyrsistent==0.15.5
pyserial==3.4
pyspark==3.5.5
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
rich==13.9.4
rich-toolkit==0.13.2
SecretStorage==2.3.1
service-identity==18.1.0
shellingham==1.5.4
simplejson==3.16.0
six==1.14.0
sniffio==1.3.1
sos==4.8.2
speedtest-cli==2.1.2
ssh-import-id==5.10
starlette==0.44.0
systemd-python==234
ttystatus==0.38
Twisted==18.9.0
typer==0.15.1
typing-extensions==4.12.2
tzdata==2025.1
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.33.0
uvloop==0.21.0
wadllib==1.3.3
watchfiles==0.24.0
websockets==13.1
wsproto==0.15.0
zipp==1.0.0
zope.interface==4.7.1
